table/requirements.txt
google-cloud-storage
google-cloud-bigquery
google-cloud-pubsub
